+++ source/extensions/omni.isaac.lab/docs/CHANGELOG.rst
@@ -1,3 +1,11 @@
+Added
+^^^^^
+
+* Added possibility to initialize cameras directly with the intrinsic matrix, which can be added as attribute to the
+  :class:`omni.isaac.lab.sim.spawner.sensors.PinholeCameraCfg` class or respectively
+  :class:`omni.isaac.lab.sensors.ray_caster.patterns_cfg.PinholeCameraPatternCfg`.
+
+
 Changelog
 ---------